+++ src/main/java/com/example/jpa/nPlusOne/entity/oneToNToNN/Three.java
@@ -1,0 +1,17 @@
+package com.example.jpa.nPlusOne.entity.oneToNToNN;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Three {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "three_id")
+    private Long id;
+
+    private String name;
+}
