+++ src/main/java/com/example/jpa/nPlusOne/repository/ARepository.java
@@ -1,6 +1,6 @@
 package com.example.jpa.nPlusOne.repository;
 
-import com.example.jpa.nPlusOne.entity.A;
+import com.example.jpa.nPlusOne.entity.multiOneToN.A;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
