+++ src/main/java/com/example/jpa/nPlusOne/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.example.jpa.nPlusOne.repository;
 
-import com.example.jpa.nPlusOne.entity.Member;
+import com.example.jpa.nPlusOne.entity.oneToN.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
